+++ scripts/install_dev_tools.sh
@@ -3,14 +3,15 @@
 # 2023-07-11 | CR
 #
 ERROR_MSG=""
-ERROR_MSG2=""
 
 INSTALL_PYENV="1"
-INSTALL_PIPENV="1"
-INSTALL_POETRY="0"
+INSTALL_PEM_TOOL="1"
+INSTALL_AWS_CLI="1"
 INSTALL_SAML2AWS="0"
 
-PYTHON_VERSION="3.9.17"
+PYTHON_VERSION=""
+
+PEM_TOOL="uv"
 
 if [ "${ERROR_MSG}" = "" ]; then
     export REPO_BASEDIR="`pwd`"
@@ -24,10 +25,11 @@ if [ "${ERROR_MSG}" = "" ]; then
 fi
 
 echo "INSTALL_PYENV=${INSTALL_PYENV}"
-echo "INSTALL_PIPENV=${INSTALL_PIPENV}"
-echo "INSTALL_POETRY=${INSTALL_POETRY}"
+echo "INSTALL_PEM_TOOL=${INSTALL_PEM_TOOL}"
 echo "INSTALL_SAML2AWS=${INSTALL_SAML2AWS}"
+echo "INSTALL_AWS_CLI=${INSTALL_AWS_CLI}"
 echo "PYTHON_VERSION=${PYTHON_VERSION}"
+echo "PEM_TOOL=${PEM_TOOL}"
 
 BASH_PROFILE="${HOME}/.bash_profile"
 BASH_SOURCE1="bash"
@@ -38,13 +40,16 @@ fi
 echo "BASH_PROFILE=${BASH_PROFILE}"
 echo "BASH_SOURCE1=${BASH_SOURCE1}"
 
-
 PYENV_INSTALL_PREFIX=""
 MACHINE_ARCHITECTURE=$(arch)
 if [[ "${MACHINE_ARCHITECTURE}" = "arm64" && "$OSTYPE" = "darwin"* ]]; then
     PYENV_INSTALL_PREFIX="arch -x86_64"
 fi
 echo "PYENV_INSTALL_PREFIX=${PYENV_INSTALL_PREFIX}"
+
+if [[ "${PYTHON_VERSION}" = "" && "${INSTALL_PYENV}" = "1" ]]; then
+    ERROR_MSG="ERROR: PYTHON_VERSION is not set"
+fi
 
 if [[ "${ERROR_MSG}" = "" && "${INSTALL_PYENV}" = "1" ]]; then
     if [ "$OSTYPE" = "linux-gnu" ]; then
@@ -177,61 +182,76 @@ if [[ "${ERROR_MSG}" = "" && "${INSTALL_PYENV}" = "1" ]]; then
     fi
 fi
 
-if [[ "${ERROR_MSG}" = "" && "${INSTALL_PIPENV}" = "1" ]]; then
-    if ! pipenv --version
-    then
-        # install pipenv
-        echo ""
-        echo "Install pipenv"
-        echo ""
-        if ! pip install --user pipenv
+if [[ "${ERROR_MSG}" = "" && "${INSTALL_PEM_TOOL}" = "1" ]]; 
+    if [ "${PEM_TOOL}" = "pipenv" ]; then
+        if ! pipenv --version
         then
-            ERROR_MSG="ERROR: pipenv could not be installed."
-        else
-            # add pipenv to the bash profile
-            ADD_CMD_TO_PROFILE="eval \"$(_PIPENV_COMPLETE=${BASH_SOURCE1}_source pipenv)\""
-            if ! grep -q "${ADD_CMD_TO_PROFILE}" ${BASH_PROFILE}; then
-                echo ""
-                echo "Add pipenv to the ${BASH_SOURCE1} profile"
-                echo ""
+            echo ""
+            echo "Install pipenv"
+            echo ""
+            if ! pip install --user pipenv
+            then
+                ERROR_MSG="ERROR: pipenv could not be installed."
+            else
+                # add pipenv to the bash profile
+                ADD_CMD_TO_PROFILE="eval \"$(_PIPENV_COMPLETE=${BASH_SOURCE1}_source pipenv)\""
+                if ! grep -q "${ADD_CMD_TO_PROFILE}" ${BASH_PROFILE}; then
+                    echo ""
+                    echo "Add pipenv to the ${BASH_SOURCE1} profile"
+                    echo ""
 
-                echo "" >> ${BASH_PROFILE}
-                echo "# `date`" >> ${BASH_PROFILE}
-                echo "# Add pipenv init" >> ${BASH_PROFILE}
-                echo "" >> ${BASH_PROFILE}
-                echo "${ADD_CMD_TO_PROFILE}" >> ${BASH_PROFILE}
+                    echo "" >> ${BASH_PROFILE}
+                    echo "# `date`" >> ${BASH_PROFILE}
+                    echo "# Add pipenv init" >> ${BASH_PROFILE}
+                    echo "" >> ${BASH_PROFILE}
+                    echo "${ADD_CMD_TO_PROFILE}" >> ${BASH_PROFILE}
+                fi
             fi
         fi
-    fi
-fi
 
-if [[ "${ERROR_MSG}" = "" && "${INSTALL_POETRY}" = "1" ]]; then
-    if ! ${HOME}/.local/bin/poetry --version
-    then
-        # install poetry
-        echo ""
-        echo "Install poetry"
-        echo ""
-        curl -sSL https://install.python-poetry.org | POETRY_HOME="${HOME}/.local" python3 -
-
+    elif [ "${PEM_TOOL}" = "poetry" ]; then
         if ! ${HOME}/.local/bin/poetry --version
         then
-            ERROR_MSG="ERROR: poetry could not be installed."
-        else
-            # add poetry to the PATH
-            if ! grep -q 'export PATH="${HOME}/.local/bin:$PATH"' ${BASH_PROFILE}; then
-                echo ""
-                echo "Add poetry to the PATH"
-                echo ""
+            echo ""
+            echo "Install poetry"
+            echo ""
+            curl -sSL https://install.python-poetry.org | POETRY_HOME="${HOME}/.local" python3 -
 
-                echo "" >> ${BASH_PROFILE}
-                echo "# `date`" >> ${BASH_PROFILE}
-                echo "" >> ${BASH_PROFILE}
-                echo "# Add poetry to the PATH" >> ${BASH_PROFILE}
-                echo "" >> ${BASH_PROFILE}
-                echo 'export PATH="${HOME}/.local/bin:$PATH"' >> ${BASH_PROFILE}
+            if ! ${HOME}/.local/bin/poetry --version
+            then
+                ERROR_MSG="ERROR: poetry could not be installed."
+            else
+                # add poetry to the PATH
+                if ! grep -q 'export PATH="${HOME}/.local/bin:$PATH"' ${BASH_PROFILE}; then
+                    echo ""
+                    echo "Add poetry to the PATH"
+                    echo ""
+
+                    echo "" >> ${BASH_PROFILE}
+                    echo "# `date`" >> ${BASH_PROFILE}
+                    echo "" >> ${BASH_PROFILE}
+                    echo "# Add poetry to the PATH" >> ${BASH_PROFILE}
+                    echo "" >> ${BASH_PROFILE}
+                    echo 'export PATH="${HOME}/.local/bin:$PATH"' >> ${BASH_PROFILE}
+                fi
             fi
         fi
+
+    elif [ "${PEM_TOOL}" = "uv" ]; then
+        if ! uv --version
+        then
+            echo ""
+            echo "Install uv"
+            echo ""
+            # On macOS and Linux.
+            curl -LsSf https://astral.sh/uv/install.sh | sh
+            if ! uv --version
+            then
+                ERROR_MSG="ERROR: uv could not be installed."
+            fi
+        fi
+    else
+        ERROR_MSG="ERROR: Invalid PEM_TOOL: ${PEM_TOOL}"
     fi
 fi
 
@@ -264,7 +284,7 @@ if [ "${ERROR_MSG}" = "" ]; then
     fi
 fi
 
-if [[ "${ERROR_MSG}" = "" && "${INSTALL_SAML2AWS}" = "1" ]]; then
+if [[ "${ERROR_MSG}" = "" && "${INSTALL_AWS_CLI}" = "1" ]]; then
     if ! aws --version
     then
         sh ${SCRIPTS_DIR}/install_aws_cli.sh
@@ -288,11 +308,7 @@ fi
 
 echo ""
 if [ "${ERROR_MSG}" = "" ]; then
-    if [ "${ERROR_MSG2}" = "" ]; then
-        echo "Done!"
-    else
-        echo "${ERROR_MSG2}" ;
-    fi
+    echo "Done!"
     echo ""
     echo "Reload the shell"
     echo ""